Change the BBR Facade to use conditional put towards Dynamo
This removes the race condition where two (or more) updates arrive in
parallel. The system will respond with an internal error.  This is
strictly not true, as it is an client error to add two different events
in parallel, but with the same sequence number.  This is not expected to
happen very often, so an internal error is fine for now.